+++ apps/bergen-meadow/app/layout.tsx
@@ -1,6 +1,8 @@
 import { GoogleAnalytics } from '@next/third-parties/google'
 import { Outfit } from 'next/font/google'
 import './globals.css'
+import Header from '../components/Header'
+import Footer from '../components/Footer'
 
 const inter = Outfit({ subsets: ['latin'] })
 
@@ -11,7 +13,11 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body className={`${inter.className} flex flex-col min-h-screen`}>
+        <Header />
+        <main className="flex-1">{children}</main>
+        <Footer />
+      </body>
       <GoogleAnalytics gaId="G-1R9Y5C3ZME" />
     </html>
   )