+++ packages/portainer-stack-redeploy/src/index.ts
@@ -9,52 +9,12 @@ import {
   Either,
 } from 'effect'
 import { equals } from 'effect/Equal'
-
-const String_UrlHost = Schema.TemplateLiteral(
-  Schema.Literal('https://', 'http://'),
-  Schema.String,
-  '.',
-  Schema.String,
-).pipe(Schema.brand('UrlHost'))
-
-const Stack = Schema.Struct({
-  Id: Schema.Number,
-  Name: Schema.String,
-  EndpointId: Schema.Number,
-})
-export type Stack = typeof Stack.Type
-
-const Stacks = Schema.Array(Stack)
-export type Stacks = typeof Stacks.Type
-
-const StackFileResponse = Schema.Struct({
-  StackFileContent: Schema.String,
-})
-export type StackFileResponse = typeof StackFileResponse.Type
-
-/**
- * Wraps the fetch API in an Effect
- */
-const Effect_fetch = (url: string | URL | Request, options?: RequestInit) =>
-  Effect.tryPromise({
-    try: () => fetch(url, options),
-    catch: (error) => {
-      console.error(error)
-      return error
-    },
-  })
-
-const Effect_responseJson = (response: Response) =>
-  Effect.tryPromise(() => response.json())
-
-const Effect_fetchJson = flow(Effect_fetch, Effect.flatMap(Effect_responseJson))
-
-const Effect_logString =
-  (message: string) =>
-  <T>(x: T) => {
-    console.log(message)
-    return Effect.succeed(x)
-  }
+import { makeApiClient } from './makeApiClient'
+import { Effect_logString } from './Effect_logString'
+import { String_UrlHost } from './String_UrlHost'
+import { Stacks } from './Stack'
+import { StackFileResponse } from './StackFileResponse'
+import { test } from './test'
 
 /**
  * @url https://github.com/wirgen/portainer-stack-redeploy-action/tree/v1.1
@@ -75,60 +35,49 @@ export const portainerStackRedeploy = async (params: {
     Either.getOrThrow,
   )
 
+  const apiClient = makeApiClient({
+    host: `${host}/api`,
+    accessToken: params.accessToken,
+  })
+
+  console.log('🔄 Getting Stack...')
+  test()
   const stack = await pipe(
-    Effect_fetchJson(`${host}/api/stacks`, {
-      method: 'GET',
-      headers: { 'X-API-Key': params.accessToken },
-    }),
+    apiClient.get('/stacks'),
+    Effect.mapError(Effect.logError),
     Effect.flatMap(Schema.decodeUnknown(Stacks)),
     Effect.flatMap(
       Array.findFirst(flow(Struct.get('Name'), equals(params.stackName))),
     ),
     Effect.runPromise,
-  )
+  ).catch((error) => {
+    console.error(error)
+    process.exit(1)
+  })
 
   console.log('🔄 Deploying Stack...')
+
   await pipe(
-    Effect_fetchJson(`${host}/api/stacks/${stack.Id}/file`, {
-      method: 'GET',
-      headers: { 'X-API-Key': params.accessToken },
-    }),
+    apiClient.get(`/stacks/${stack.Id}/file`),
     Effect.flatMap(Schema.decodeUnknown(StackFileResponse)),
     Effect.map(Struct.get('StackFileContent')),
     Effect.tap(Effect_logString('💾 Updating Stack...')),
     Effect.flatMap((stackFile) =>
-      Effect_fetchJson(
-        `${host}/api/stacks/${stack.Id}?endpointId=${stack.EndpointId}`,
-        {
-          method: 'PUT',
-          headers: { 'X-API-Key': params.accessToken },
-          body: JSON.stringify({
-            stackFileContent: stackFile,
-            pullImage: true,
-          }),
-        },
-      ),
+      apiClient.put(`/stacks/${stack.Id}?endpointId=${stack.EndpointId}`, {
+        stackFileContent: stackFile,
+        pullImage: true,
+      }),
     ),
     Effect.tap(Effect_logString('💾 Stack updated')),
     Effect.tap(Effect_logString('✋ Stopping Stack...')),
     Effect.flatMap(() =>
-      Effect_fetchJson(
-        `${host}/api/stacks/${stack.Id}/stop?endpointId=${stack.EndpointId}`,
-        {
-          method: 'POST',
-          headers: { 'X-API-Key': params.accessToken },
-        },
-      ),
+      apiClient.post(`/stacks/${stack.Id}/stop?endpointId=${stack.EndpointId}`),
     ),
     Effect.tap(Effect_logString('✋ Stack stopped')),
     Effect.tap(Effect_logString('🚀 Starting Stack...')),
     Effect.flatMap(() =>
-      Effect_fetchJson(
-        `${host}/api/stacks/${stack.Id}/start?endpointId=${stack.EndpointId}`,
-        {
-          method: 'POST',
-          headers: { 'X-API-Key': params.accessToken },
-        },
+      apiClient.post(
+        `/stacks/${stack.Id}/start?endpointId=${stack.EndpointId}`,
       ),
     ),
     Effect.tap(Effect_logString('🚀 Stack deployed')),