+++ apps/ally-personal/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @repo/ally-personal
 
+## 1.2.8
+
+### Patch Changes
+
+- 7a4fc16: ally personal tweaks post board member
+
+## 1.2.7
+
+### Patch Changes
+
+- 1d2fdb1: ally personal bergen meadow update tab
+- 951a3df: fix ally bm image sizes and name
+- db436dd: Added bergen meadow update tab
+
+## 1.2.6
+
+### Patch Changes
+
+- 570843e: Added new forum page, added latest change to footer
+
 ## 1.2.5
 
 ### Patch Changes